rag_chatbot/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.52.2
qdrant-client==1.12.1
sqlalchemy==2.0.35
python-multipart==0.0.12
sentence-transformers==3.0.1
torch
transformers